adding new folder, utils folder, to put the computeEndDate function
