+++ frontend/src/components/code-connect/FormCreate.tsx
@@ -10,6 +10,7 @@ import {
   contentTechsBackCodeConnect,
   contentTechsFrontCodeConnect,
 } from "./techsLabelsContent";
+import { computeEndDate } from "../code-connect/utils/computeEndDate";
 
 const getMinDeadline = () => {
   const tomorrow = new Date();
@@ -46,22 +47,6 @@ const FormCreate = () => {
         end_date: computeEndDate(prev.start_date ?? "", prev.time, value),
       }),
     }));
-  };
-
-  const computeEndDate = (
-    startDate: string,
-    time: number,
-    unitTime: string,
-  ): string => {
-    if (!startDate || !time || !unitTime) return "";
-    const [year, month, day] = startDate.split("-").map(Number);
-    const date = new Date(year, month - 1, day); // local time, not UTC
-    if (unitTime === "week") date.setDate(date.getDate() + time * 7);
-    if (unitTime === "month") date.setMonth(date.getMonth() + time);
-    const y = date.getFullYear();
-    const m = String(date.getMonth() + 1).padStart(2, "0");
-    const d = String(date.getDate()).padStart(2, "0");
-    return `${y}-${m}-${d}`;
   };
 
   const handleInputsNumber = (
